Day 1
This completes part 1 and part 2 of the Advent of Code 2020.

This is basically an example of me stumbling through a bunch of Go
tutorials in order to read files, convert it to integers and then do
some bubble sort logic on the results.

There are vastly superior ways to complete this challenge and I'm likely
to come back to them. For now though, happy to complete day one.